+++ src/main/java/net/sf/mpxj/primavera/PrimaveraReader.java
@@ -188,6 +188,7 @@ final class PrimaveraReader
          properties.setFiscalYearStartMonth(row.getInteger("fy_start_month_num"));
          properties.setUniqueID(row.getInteger("proj_id"));
          properties.setExportFlag(row.getBoolean("export_flag"));
+         properties.setMustFinishBy(row.getDate("scd_end_date"));         
          // cannot assign actual calendar yet as it has not been read yet
 
          m_defaultCalendarID = row.getInteger("clndr_id");