+++ src/main/java/net/sf/mpxj/primavera/PrimaveraXERFileReader.java
@@ -1183,6 +1183,7 @@ public final class PrimaveraXERFileReader extends AbstractProjectStreamReader
       FIELD_TYPE_MAP.put("resume_date", XerFieldType.DATE);
       FIELD_TYPE_MAP.put("role_id", XerFieldType.INTEGER);
       FIELD_TYPE_MAP.put("rsrc_id", XerFieldType.INTEGER);
+      FIELD_TYPE_MAP.put("scd_end_date", XerFieldType.DATE);
       FIELD_TYPE_MAP.put("sched_calendar_on_relationship_lag", XerFieldType.STRING);
       FIELD_TYPE_MAP.put("seq_num", XerFieldType.INTEGER);
       FIELD_TYPE_MAP.put("start_date", XerFieldType.DATE);