+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 9.5.2 (git master)
+* Add Must Finish By date to project properties.
 
 ## 9.5.1 (01/07/2021)
 * When applying a baseline using ProjectFile.setBaseline, gracefully handle duplicate task key values.